[[Imported from SVN: r6564]]

+++ dune/grid/common/test/test-genericquadrature.cc
@@ -1,5 +1,6 @@
-// -*- tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 2 -*-
-// vi: set et ts=4 sw=2 sts=2:
+// NOTE: The current revision of this file was left untouched when the DUNE source files were reindented!
+// NOTE: It contained invalid syntax that could not be processed by uncrustify.
+
 #include <limits>
 #include <iostream>
 
@@ -13,13 +14,13 @@
 bool success = true;
 
 /*
-   This is a simple accuracy test on the reference element. It integrates
-   x^p and y^p with the quadrature rule of order p, which should give
-   an exact result.
+  This is a simple accuracy test on the reference element. It integrates
+  x^p and y^p with the quadrature rule of order p, which should give
+  an exact result.
  */
 
 /*
-   Exact (analytical) solution on different reference elements.
+  Exact (analytical) solution on different reference elements.
  */
 
 template <class ctype, int dim>
@@ -29,11 +30,11 @@ ctype analyticalSolution (Dune::GeometryType t, int p, int direction )
   ctype exact = 0;
   switch (t.basicType())
   {
-  case GeometryType::cube :
+  case GeometryType::cube:
     exact=1.0/(p+1);
     break;
 
-  case GeometryType::simplex :
+  case GeometryType::simplex:
     /* 1/(prod(k=1..dim,(p+k)) */
     exact = ctype( 1 );
     for( int k = 1; k <= dim; ++k )
@@ -41,7 +42,7 @@ ctype analyticalSolution (Dune::GeometryType t, int p, int direction )
     exact = ctype( 1 ) / exact;
     break;
 
-  case GeometryType::prism :
+  case GeometryType::prism:
   {
     const int pdim = (dim > 0 ? dim-1 : 0);
     if( direction < dim-1 )
@@ -57,19 +58,19 @@ ctype analyticalSolution (Dune::GeometryType t, int p, int direction )
     break;
   }
 
-  case GeometryType::pyramid :
+  case GeometryType::pyramid:
     switch( direction )
     {
-    case 0 :
-    case 1 :
+    case 0:
+    case 1:
       exact=1.0/((p+3)*(p+1));
       break;
-    case 2 :
+    case 2:
       exact=2.0/((p+1)*(p+2)*(p+3));
       break;
     };
     break;
-  default :
+  default:
     DUNE_THROW(Dune::NotImplemented, __func__ << " for " << t);
   };
   return exact;
@@ -91,23 +92,23 @@ void checkQuadrature(const Quadrature &quad)
     const ctype weight = qp->weight();
 
     for (unsigned int d=0; d<dim; d++)
-      integral[d] += weight*std::pow(x[d],p);
+      integral[d] += weight*std::pow(x[d],double(p);
   }
-
+  
   ctype maxRelativeError = 0;
   int dir = -1;
   for( unsigned int d=0; d<dim; d++ )
   {
     ctype exact = analyticalSolution<ctype,dim>(t,p,d);
     ctype relativeError = std::abs(integral[d]-exact) /
-                          (std::abs(integral[d])+std::abs(exact));
+      (std::abs(integral[d])+std::abs(exact));
     if (relativeError > maxRelativeError)
     {
       maxRelativeError = relativeError;
       dir = d;
     }
   }
-  ctype epsilon = std::pow(2.0,p)*p*std::numeric_limits<double>::epsilon();
+  ctype epsilon = std::pow(2.0,double(p))*p*std::numeric_limits<double>::epsilon();
   if (p==0)
     epsilon = 2.0*std::numeric_limits<double>::epsilon();
   if (maxRelativeError > epsilon) {
@@ -116,12 +117,12 @@ void checkQuadrature(const Quadrature &quad)
     {
       ctype exact = analyticalSolution<ctype,dim>(t,p,d);
       ctype relativeError = std::abs(integral[d]-exact) /
-                            (std::abs(integral[d])+std::abs(exact));
+        (std::abs(integral[d])+std::abs(exact));
       std::cerr << "       relative error " << relativeError << " in direction " << d << " (exact = " << exact << " numerical = " << integral[d] << ")" << std::endl;
     }
     success = false;
   }
-}
+}  
 
 template<class Quadrature>
 void checkWeights(const Quadrature &quad)
@@ -140,14 +141,14 @@ void checkWeights(const Quadrature &quad)
   }
   if (std::abs(volume -
                Dune::GenericReferenceElements<ctype, dim>::general(t).volume())
-      > 4*dim*(p ? p : 1)*std::numeric_limits<double>::epsilon())
+      > 4*dim*(p?p:1)*std::numeric_limits<double>::epsilon())
   {
     std::cerr << "Error: Quadrature for " << t << " and order=" << p
               << " does not sum to volume of RefElem" << std::endl;
     std::cerr << "\tSums to " << volume << "( RefElem.volume() = "
               << Dune::GenericReferenceElements<ctype, dim>::general(t).volume()
               << ")" << "(difference " << volume -
-    Dune::GenericReferenceElements<ctype, dim>::general(t).volume()
+      Dune::GenericReferenceElements<ctype, dim>::general(t).volume()
               << ")" << std::endl;
     success = false;
   }
@@ -159,17 +160,17 @@ void check( const Dune::GeometryType::BasicType &btype, unsigned int maxOrder )
   typedef Dune::GenericGeometry::GaussPoints<CF> OneDPoints;
   typedef Dune::GenericGeometry::GenericQuadratureFactory<dim,double,OneDPoints> QuadratureProvider;
   typedef typename QuadratureProvider::Object Quadrature;
-  for (unsigned int p=0; p<=maxOrder; ++p)
+  for (unsigned int p=0;p<=maxOrder; ++p)
   {
     const Quadrature &quad = *QuadratureProvider::create(Dune::GeometryType(btype,dim),p);
     checkWeights(quad);
     checkQuadrature(quad);
     QuadratureProvider::release(&quad);
   }
-  if (dim>0 && (dim>2 ||
-                btype==Dune::GeometryType::cube ||
-                btype==Dune::GeometryType::simplex) )
-    check<CF,(dim==0) ? 0 : dim-1>(btype,maxOrder);
+  if (dim>0 && (dim>2 || 
+               btype==Dune::GeometryType::cube ||
+               btype==Dune::GeometryType::simplex) )
+    check<CF,(dim==0)?0:dim-1>(btype,maxOrder);
 }
 
 int main ()
@@ -179,7 +180,7 @@ int main ()
     check<double,4>(Dune::GeometryType::simplex,55);
     check<double,4>(Dune::GeometryType::prism,55);
     check<double,3>(Dune::GeometryType::pyramid,55);
-  }
+ }
   catch( const Dune::Exception &e )
   {
     std::cerr << e << std::endl;
@@ -189,6 +190,6 @@ int main ()
     std::cerr << "Generic exception!" << std::endl;
     return 1;
   }
-
-  return success ? 0 : 1;
+  
+  return success ? 0:1;
 }
